+++ data.go
@@ -2,5 +2,5 @@ package uadmin
 
 type listData struct {
 	Rows  [][]interface{}
-	Count int64
+	Count int
 }
